Move dataset functions outside of data/__init__
