+++ kikuchipy/indexing/_orientation_similarity_map.py
